CI: Add workflow to update package-lock.json every month
Adds a workflow that generates a up to date lockfile based on the current package.json. The workflow runs every 1st day of each month, and if either this configuration or the package.json itself is modified. A run will open a PR with the updated package-lock.json, and if a PR is already open it will update that PR.

By running and opening/updating a PR each month it's ensured the package-lock.json stays up to date and updated frequently at convenient times, while reproducible builds are still ensured.