Ungzip source tarballs manually
When using node-fetch, NASM website transfers *.tar.gz with HTTP headers
for transfer compression and node-fetch ungzips it on the fly, but when
GitHub mirror is used, the file is downloaded verbatim and needs to be
decompressed. In order to avoid dealing with this, tell node-fetch to
download the file as is, we'll decompress it later manually.